Refactored a bit the rds module
It now requires the module user to be a bit more explicit
when specifying engines and engine versions.
It previously assumed some parameters like engine versions and
parameter group families according to the provided engine, and it was
too cumbersome to override those.
Also changed parameter name `rds_type` to `engine`, which seemed more
appropriate.

Will bump major version for this.