Change the snapshot period variable format
To avoid manual mistakes when setting up the monitoring period, we'll use a simple rate expression to set the snapshot period, in hours, and automatically set the monitoring period to the same value (plus 1 hour).

Before the monitoring period could easily be missed, and if set to something lower than the snapshot period, could cause missed alerts.