Fix a bug with for aurora clusters (#63)
With the removal of the "replicating" tag, now the step3 lambda function was trying to copy all matching snapshots to the target account every time it ran (every 30 minutes), which caused problems when the snapshot had already been copied. This now checks if the destination snapshot already exists and just moves on if so.

This isn't a problem with non-cluster snapshots since the step3 lambda function is triggered via a snapshot event, instead of a cron.